Fix static_cast that triggers compile errors with -Wpedantic (#1632)
(cherry picked from commit 645dd5485b2c35ea2a5a7b3433c838585e6c5972)

+++ src/ParticleEmitter.cc
@@ -267,7 +267,7 @@ bool ParticleEmitter::SetType(const std::string &_typeStr)
 /////////////////////////////////////////////////
 std::string ParticleEmitter::TypeStr() const
 {
-  size_t index = static_cast<int>(this->dataPtr->type);
+  size_t index = static_cast<size_t>(this->dataPtr->type);
   if (index < kEmitterTypeStrs.size())
     return std::string(kEmitterTypeStrs[index]);
   return "point";
